Basically its a Secure Online voting System Which can be used in Schools and Colleges To Vote for their favourite Candidate
Admin has full control of the system, all the functions are to be performed from Admin panel. Here, the user can view total votes with respective voters name, add and list voters, positions, candidates, view ballot positions, edit election title and edit admin’s profile. To vote for an election, you need a voter and to add a voter, the user has to provide first and last name, password and upload a photo. Likewise, to add a position a short description and the maximum vote value should be provided. All the candidates can be views from the admin panel with their respective positions and similarly to add candidates for an election the user should provide a first name and last name, select position, upload a photo and write the platform.

Ballot positions can be maintained for the voters according to the admin’s will and the admin can change the election title easily anytime he/she wants. After all the votings, total votes can be viewed from Votes navigation, which displays Position with candidate and voter name. Before logging in as a voter, the user needs a voter account and a voter id. That can be retrieved from Voters List, the system generates Voters id automatically and the user has to copy it then paste in voter’s login section and provide a password to continue as a voter. A clean and responsive dashboard is provided in the admin panel for the easy management of the voting system. Voting system in PHP helps in easy management of the votes for certain objectives. Design of this project is pretty and responsive so that user won’t find it difficult to understand, use and navigate.

Voting System project is developed using PHP, JavaScript, Bootstrap, and CSS. Talking about the project, it has lots of essential features. This project contains a Voter’s login side where a voter can Sign in to vote and Admin Panel where he/she can view total votes, add and list voters, positions, candidates and many more. While logging in from voter’s login, the user should provide Voters ID in order to log in to the system to vote. The voter’s id can be retrieved from the Admin Panel.